refactor(core): reorganize systems/ into domain subdirectories
Group the 28 flat files in app/core/src/systems/ into five subdirectories:
- enemies/  (ai, cull, difficulty, spawn)
- player/   (mod, collision)
- projectiles/ (mod, collision)
- weapons/  (bible, cooldown, cross, fire_wand, garlic, knife, magic_wand, thunder_ring, whip)
- xp/       (apply, attraction, choices, drop, level_up)

Pure rename/restructure — no logic changes. All 320 tests pass.
