+++ app/core/src/events.rs
@@ -89,7 +89,7 @@ pub struct GameOverEvent;
 
 /// Fired when accumulated XP reaches the next-level threshold.
 ///
-/// The [`check_level_up`](crate::systems::level_up::check_level_up) system
+/// The [`check_level_up`](crate::systems::xp::level_up::check_level_up) system
 /// emits this event and transitions to
 /// [`AppState::LevelUp`](crate::states::AppState::LevelUp) so the player can
 /// choose an upgrade card.
