+++ app/core/src/systems/enemies/cull.rs
