+++ app/core/src/systems/weapons/mod.rs
@@ -1,0 +1,9 @@
+pub mod bible;
+pub mod cooldown;
+pub mod cross;
+pub mod fire_wand;
+pub mod garlic;
+pub mod knife;
+pub mod magic_wand;
+pub mod thunder_ring;
+pub mod whip;
