+++ app/core/src/systems/xp/apply.rs
