+++ app/core/src/systems/xp/attraction.rs
