+++ app/core/src/systems/xp/choices.rs
