+++ app/core/src/systems/xp/mod.rs
@@ -1,0 +1,5 @@
+pub mod apply;
+pub mod attraction;
+pub mod choices;
+pub mod drop;
+pub mod level_up;
